+++ mrp_account_analytic_wip/__manifest__.py
@@ -10,8 +10,12 @@
     "author": "Open Source Integrators, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/manufacture",
     "license": "AGPL-3",
-    "depends": ["account_analytic_wip", "mrp_analytic_cost"],
-    "data": ["views/mrp_production_views.xml"],
+    "depends": ["account_analytic_wip", "mrp"],
+    "data": [
+        "views/account_analytic_line_view.xml",
+        "views/mrp_production_views.xml",
+        "views/mrp_workcenter_view.xml",
+    ],
     "installable": True,
     "maintainers": ["dreispt"],
     "development_status": "Alpha",
