+++ mrp_account_analytic_wip/models/__init__.py
@@ -1,3 +1,4 @@
+from . import account_analytic_line
 from . import account_analytic_tracking
 from . import mrp_production
 from . import mrp_workorder
