+++ mrp_account_analytic_wip/readme/CONFIGURATION.rst
@@ -1,3 +1,7 @@
+* Create a Product for each cost type needed. Set the standard cost to use for each of them.
+  It is also recommended to properly organize them in Product Categories.
+* On each Work Center, select the Cost Type Products to use for the operrations generated Analytic Items.
+
 On the Product Category, "Operations Costing" section, configure the accounts to use:
 
 * Consumption Account: the account to credit.
