+++ mrp_account_analytic_wip/readme/USAGE.rst
@@ -1,4 +1,12 @@
 To use:
 
 * On the Manufacturing Order, set the Analytic Account to use. This may correspond to a Project.
-* At Manufacturing > Work Orders, track time spent on operations, as usual. These records will automatically generate cost records as Analytic Items.
+* On Manufacturing Orders, Analytic Items are automatically generated when:
+
+  * Raw materials are consumed, or
+  * Time is spent on work order Operations, with a Work Center with a Cost Type product set.
+
+To analyze costs:
+
+* Go to Manufacturing > Reports > Analytic Items. The cost data is available there and can be used for analysis.
+* At Project > Dashboard, "Project Overview" form, a summary of the costs for the corresponding Analytic Account is presented.
