+++ mrp_analytic_cost/__init__.py
@@ -1,4 +1,2 @@
 # Copyright (C) 2020 Open Source Integrators
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import models
