+++ mrp_analytic_cost/__manifest__.py
@@ -9,11 +9,6 @@
     "author": "Open Source Integrators, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/manufacture",
     "license": "AGPL-3",
-    "depends": ["mrp_analytic", "analytic_activity_based_cost"],
-    "data": [
-        "views/account_analytic_line_view.xml",
-        "views/mrp_workcenter_view.xml",
-    ],
     "maintainer": "dreispt",
     "development_status": "Alpha",
     "installable": True,
