--- mrp_analytic_cost/i18n/hu.po
@@ -1,127 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* mrp_analytic_cost_material
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: Automatically generated\n"
-"Language-Team: none\n"
-"Language: hu\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=2; plural=n != 1;\n"
-
-#. module: mrp_analytic_cost
-#: model:ir.actions.act_window,name:mrp_analytic_cost.action_new_mrp_analytic_items
-#: model:ir.ui.menu,name:mrp_analytic_cost.custom_account_analytic_line_menu
-msgid "Analytic Items"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model,name:mrp_analytic_cost.model_account_analytic_line
-msgid "Analytic Line"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_workcenter__analytic_product_id
-msgid "Cost Type"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_workcenter__costs_hour
-msgid "Cost per hour"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_account_analytic_line__display_name
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_production__display_name
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_workcenter__display_name
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_workorder__display_name
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_stock_move__display_name
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_stock_move_line__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_account_analytic_line__id
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_production__id
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_workcenter__id
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_workorder__id
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_stock_move__id
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_stock_move_line__id
-msgid "ID"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_account_analytic_line____last_update
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_production____last_update
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_workcenter____last_update
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_mrp_workorder____last_update
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_stock_move____last_update
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_stock_move_line____last_update
-msgid "Last Modified on"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model_terms:ir.ui.view,arch_db:mrp_analytic_cost.custom_account_analytic_line_form_view
-msgid "Manufacture"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model_terms:ir.ui.view,arch_db:mrp_analytic_cost.custom_account_analytic_line_filter_view
-msgid "Manufacturing Order"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model_terms:ir.ui.view,arch_db:mrp_analytic_cost.custom_mrp_production_form_view
-msgid "Material Costs"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model,name:mrp_analytic_cost.model_stock_move_line
-msgid "Product Moves (Stock Move Line)"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model,name:mrp_analytic_cost.model_mrp_production
-msgid "Production Order"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_account_analytic_line__manufacturing_order_id
-msgid "Related Manufacturing Order"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_account_analytic_line__stock_move_id
-msgid "Related Stock Move"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model_terms:ir.ui.view,arch_db:mrp_analytic_cost.custom_account_analytic_line_filter_view
-msgid "Related to MO"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model.fields,help:mrp_analytic_cost.field_mrp_workcenter__costs_hour
-msgid "Specify cost of work center per hour."
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model,name:mrp_analytic_cost.model_stock_move
-msgid "Stock Move"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model,name:mrp_analytic_cost.model_mrp_workcenter
-msgid "Work Center"
-msgstr ""
-
-#. module: mrp_analytic_cost
-#: model:ir.model,name:mrp_analytic_cost.model_mrp_workorder
-#: model:ir.model.fields,field_description:mrp_analytic_cost.field_account_analytic_line__workorder_id
-msgid "Work Order"
-msgstr ""
