--- mrp_analytic_cost/models/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (C) 2020 Open Source Integrators
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import account_analytic_line
-from . import mrp_production
-from . import mrp_workorder
-from . import stock_move
