--- mrp_analytic_cost/readme/CONFIGURATION.rst
@@ -1,3 +1,0 @@
-* Create a Product for each cost type needed. Set the standard cost to use for each of them.
-  It is also recommended to properly organize them in Product Categories.
-* On each Work Center, select the Cost Type Products to use for the operrations generated Analytic Items.
