+++ mrp_analytic_cost/readme/DESCRIPTION.rst
@@ -1,12 +1,1 @@
-Out of the box, manufacturing costs are only generated
-when a Manufacturing Order is closed.
-
-There are cases where the manufacturing costs need to be tracked in real time.
-This is especially important for manufacturing scenarios with a long cycle time,
-spanning for multiple months.
-
-This feature allows to track costs while the manufacturing is in progress,
-both for raw materials used and work order operations.
-
-Costs incurred are stored as Analytic Items, and can then be analyzed.
-No journal entries are generated.
+DEPRECATED: this feature was merged into mrp_account_analytic_wip
