+++ mrp_analytic_cost/readme/USAGE.rst
@@ -1,12 +1,1 @@
-To use:
-
-* On the Manufacturing Order, set the Analytic Account to use. This may correspond to a Project.
-* On Manufacturing Orders, Analytic Items are automatically generated when:
-
-  * Raw materials are consumed, or
-  * Time is spent on work order Operations, with a Work Center with a Cost Type product set.
-
-To analyze costs:
-
-* Go to Manufacturing > Reports > Analytic Items. The cost data is available there and can be used for analysis.
-* At Project > Dashboard, "Project Overview" form, a summary of the costs for the corresponding Analytic Account is presented.
+DEPRECATED: this feature was merged into mrp_account_analytic_wip
